eliminate `chia.types.spend_bundle`
